Rename file board-controller -> board
